+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 Version 0.16.0 *(In development)*
 ---------------------------------
 
+- Add `pomFromGradleProperties` API to base plugin. This configures the pom in the same way the regular plugin does.
 - The deprecated `nexusOptions` method was removed. `closeAndReleaseRepository` is automatically configured.
+
 
 Version 0.15.1 *(2021-05-02)*
 ---------------------------------